fastapi_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai>=1.0.0
